+++ packages/admin/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@awesome-comment/admin",
-  "version": "0.2.5",
+  "version": "0.3.0",
   "private": true,
   "scripts": {
     "build": "nuxt build",
@@ -17,7 +17,7 @@
     "autoprefixer": "^10.4.15",
     "daisyui": "^3.7.5",
     "nuxt": "^3.7.3",
-    "postcss": "^8.4.29",
+    "postcss": "^8.4.30",
     "pug": "^3.0.2",
     "tailwindcss": "^3.3.3"
   },
