+++ packages/admin/pages/admin/login.vue
@@ -1,0 +1,13 @@
+<script lang="ts" setup>
+
+</script>
+
+<template lang="pug">
+
+</template>
+
+<script lang="ts">
+export default {
+  name: 'AdminLoginPage',
+}
+</script>
