+++ packages/admin/server/middleware/auth.ts
@@ -1,0 +1,9 @@
+export default defineEventHandler(async function (event) {
+  // not visit admin
+  const url = getRequestURL(event);
+  if (!url.pathname.startsWith('/admin')) {
+    return;
+  }
+
+  event.context.auth = await checkUserPermission(event);
+});
