+++ packages/admin/types/index.ts
@@ -1,0 +1,12 @@
+export type AcConfig = {
+  adminEmails: string[];
+}
+
+export type ApiResponse<T> = {
+  code: number;
+  data?: T;
+  message?: string;
+  meta?: {
+    total: number;
+  };
+}
